Issue #20 Update sprite current_location_x/y to unsigned int.
This needs to be greater than a byte as it must cover the entire map (not
just the entire map in the view of the gameboy, which can be an unsigned
int).
-e 
Original/build commit ID: 0e76730ca60a4ae67ad1ec454b1622ba32d377c7

+++ game_state.c
@@ -80,8 +80,8 @@ typedef struct {
     // The direction is inverted when max/min location is reached.
     INT8 travel_direction_x;
     INT8 travel_direction_y;
-    UINT8 current_location_x;
-    UINT8 current_location_y;
+    unsigned int current_location_x;
+    unsigned int current_location_y;
     unsigned int min_location;
     unsigned int max_location;
 } ai_sprite;
